20/11/2018-0 - adds progress bar for adding cities to the database - minor corrections

+++ MyWeatherApp/WeatherModels/Main.cs
@@ -4,6 +4,6 @@ namespace MyWeatherApp.WeatherModels
     {
         public float Temp { get; set; }
         public float Pressure { get; set; }
-        public float Humidity { get; set; }
+        public int Humidity { get; set; }
     }
 }
